move AddressFromMap to internal
